requirements.txt
boto3==1.26.78
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
filelock==3.8.0
Flask==2.2.2
grpcio==1.47.2
grpcio-tools==1.47.2
huggingface-hub==0.11.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kafka-python==2.0.2
kazoo==2.5.0
keras-facenet==0.3.2
MarkupSafe==2.1.1
mmh3==3.0.0
nltk==3.7
numpy==1.23.5
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
protobuf==3.20.3
pymilvus==2.1.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
tabulate==0.9.0
tensorflow==2.10.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
ujson==5.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.10.0
minio==7.1.13
